todo item: Move realm(s) to database (#787)

+++ main.py
@@ -1,6 +1,7 @@
 import multiprocessing
 import os
-from sys import platform
+import argparse
+from sys import platform, argv
 from time import sleep
 
 from game.realm import RealmManager
@@ -13,6 +14,10 @@ from utils.ChatLogManager import ChatLogManager
 from utils.PathManager import PathManager
 from utils.constants import EnvVars
 
+#Argument Parser
+parser = argparse.ArgumentParser(description="Alpha-Core help", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+parser.add_argument("-l", help="Usage: -l realm -l world", dest='launch', action='store_true')
+args, unknown = parser.parse_known_args()
 
 def release_process(process):
     while process.is_alive():
@@ -61,15 +66,27 @@ if __name__ == '__main__':
         if env_var:
             Logger.info(f'Environment variable {env_var_name}: {env_var}')
 
-    login_process = context.Process(target=RealmManager.LoginServerSessionHandler.start)
-    login_process.start()
+    if not args.launch:
+        login_process = context.Process(target=RealmManager.LoginServerSessionHandler.start)
+        login_process.start()
 
-    proxy_process = context.Process(target=RealmManager.ProxyServerSessionHandler.start)
-    proxy_process.start()
+        proxy_process = context.Process(target=RealmManager.ProxyServerSessionHandler.start)
+        proxy_process.start()
 
-    world_process = context.Process(target=WorldManager.WorldServerSessionHandler.start)
-    world_process.start()
-
+        world_process = context.Process(target=WorldManager.WorldServerSessionHandler.start)
+        world_process.start()
+    else:
+        if (argv[-1] == "realm"):
+            login_process = context.Process(target=RealmManager.LoginServerSessionHandler.start)
+            login_process.start()
+        
+            proxy_process = context.Process(target=RealmManager.ProxyServerSessionHandler.start)
+            proxy_process.start()
+        
+        if (argv[-1] == "world"):
+            world_process = context.Process(target=WorldManager.WorldServerSessionHandler.start)
+            world_process.start()
+        
     # noinspection PyBroadException
     try:
         if os.getenv('CONSOLE_MODE', config.Server.Settings.console_mode) in [True, 'True', 'true']:
@@ -83,17 +100,36 @@ if __name__ == '__main__':
     ChatLogManager.exit()
     
     # Send SIGTERM to processes.
-    world_process.terminate()
-    Logger.info('World process terminated.')
-    proxy_process.terminate()
-    Logger.info('Proxy process terminated.')
-    login_process.terminate()
-    Logger.info('Login process terminated.')
+    # Add checks to send SIGTERM to only running process
+    if argv:
+        if argv[-1] == "world":
+            world_process.terminate()
+            Logger.info('World process terminated.')
+        if argv[-1] == "realm":
+            login_process.terminate()
+            Logger.info('Login process terminated.')
+            proxy_process.terminate()
+            Logger.info('Proxy process terminated.')
+    else:
+        world_process.terminate()
+        Logger.info('World process terminated.')
+        proxy_process.terminate()
+        Logger.info('Proxy process terminated.')
+        login_process.terminate()
+        Logger.info('Login process terminated.')
 
     # Release process resources.
     Logger.info('Waiting to release resources...')
-    release_process(world_process)
-    release_process(proxy_process)
-    release_process(login_process)
+    
+    #These checks are needed to prevent an end of program crash with keyboard interrupt
+    if argv[-1] == "world":
+        release_process(world_process)
+    elif argv[-1] == "realm":
+        release_process(proxy_process)
+        release_process(login_process)
+    else:
+        release_process(world_process)
+        release_process(proxy_process)
+        release_process(login_process)
 
     Logger.success('Core gracefully shut down.')
